Fixed loading of collections including all their instances.
This required to changed collections to use the global instance store
implemented in dlite-entity.c instead of having its own.

Running from Python seems to work well, with no definite memory leaks,
but the test_collection test now reports a memory leak.  Seems to be a
bug in the test itself - but it should be fixed anyway...