Fixed a range of VS compilation warnings
A major issue on Windows is that we cannot provide correct paths,
since path separators on Windows are semicolon which cmake
add_custom_command() converts into spaces.  There seems no way to
prevent that - escaping with semicolon worked a few cases but not
consistently.  Trying out to make a small env-like tool that runs
an executable with a pre-defined dlite environment.