chemistry-reaction.ttl to chemistry.ttl and MoleculeModel
Changed name of ontology to chemistry.ttl so that it does sound
relevant for both 'calculations'
Added MoleculeModel with hasProperty PotentialEnergy.
This should be changed at some later stage. As long as it
has any energy that is a state function it can be used to calculate
reaction energy, but they have to be the same for all participating
reactants and products. Also deltaStateFunctions can be used,
provided the references are correct (simplet example, standard
energies of formation).