Changed mapsTo from rdf:Property to owl:AnnotationProperty.
Protege and Owlready2 (and therefore also EMMOntoPy) are based on owl as
base ontology and they change the ontology when loading the original
mapsTo. This is not very practical. We therefore changed the ontlogy
instad of updating EMMOntoPy. This might be changed in the future.
Commented in the README.md