Updated cppcheck and corrected some errors that it found (#911)
Updated cppcheck and corrected some errors that it found.

Background for this PR: After updating cppcheck to 2.13.0, it pointed out a few memory leaks. To make it easier to utilise cppcheck, this PR runs it with cache enabled. After the initial run, this makes cppcheck very fast, making it convenient to use during development.

Co-authored-by: Francesca L. Bleken <48128015+francescalb@users.noreply.github.com>