Include the Dockerfile we build in the sha hash for checking whether a new image should be published (#1023)
The current implementation of the container_builds_weekly.yml only publishes the build container if the image on quay.io/pypa/ has changed, not if we have made any changes in the Dockerfiles that we prepare for building DLite. This means that any changes we do in the Dockerfiles will not be available for the ci_build_wheels.yml action.

This PR fixes this broken behaviour, by also including the hash of our local Dockerfile in the file for comparing hashes.