Improved calculation of IRIs (#1112)
Updates the way IRIs are calculated as suggested in the update of issue #788.

In addition the PR includes some bug fixes and documentation cleanup. It turned out that un-namespaced ids were used in many tests, which had to be updated. Hence the PR ended up being quite large.

Added dlite_info() and dlite_debug() functions.